Install script changes to work out of the box on OpenShift 4
These are the tweaks I had to make to get `etc/scripts/install.sh`
working on OpenShift 4. I haven't manually verified everything yet,
but this does get Maistra and all 3 parts of Knative installed.

To fix the duplicate Image CRD problem I just deleted its definition
from Serving. That obviously is a problem if Serving is used without
Build, but for the purposes of our install script that's not the case.