More updates for OpenShift 4
Create OperatorGroups for each Operator before we create the
Subscription for that Operator. It does eventually work doing it the
other way around but this gets things started faster.

Those OperatorGroups specify a single target namespace that exactly
matches the namespace each component gets installed into.

Remove the olm.* annotations from the ClusterServiceVersions - these
are meant to be set by OLM and weren't only because we weren't using
OLM quite right for our scenario.

Add a few StatusDescriptors for some of our custom resources so the
user gets a bit more information when browsing them via the web
console.