Extra RBAC rules for kubecon demo + OS 3.11 + knative 0.3 [OCF-377]
This commit shows the gap from upstream's RBAC policies and ones
needed to run on OpenShift. Previous CSV's were grossly assigning
cluster-admin privs, but as of 0.3.0 no longer, except for eventing
since upstream uses cluster-admin privs currently.

It's curious that build needed a rule for 'services' and eventing
needed a wildcard for its finalizers, since I would've expected that
to be a problem even on vanilla k8s. And the rule for
'podsecuritypolicies' seems self-imposed due to our initContainer
hack: the upstream manifest relies on ordering to avoid that.