Put the hack on the controller instead of the activator
Ideally, this would be in its own operator deployment, but right now
that image isn't allowed to terminate: 'Always' is the only restart
policy allowed.

At least, on the controller, we avoid the more likely possibility of
many olm-hacks firing as a result of many activators responding to
load. They're harmless, but also useless.

We need "real" operators.