Reverting back to known OLM sha and watching deployments
It's no longer enough to wait for all the pods in a namespace since
the operator is running there. So we wait on a deployment first to
ensure we have enough pods to wait for, if that makes any sense.

While I was at it, I simplified the timeout logic to wait for success
0 instead of error !0, which was weird.

Using OLM's later than early-febururrurary-ish result in errors trying
to deploy knative-eventing referring to label names longer than 63
characters.

Whew!